requirements.txt
loguru
mkdocs
pip
python-dotenv
ruff
tqdm
typer
gdown
opencv-python
matplotlib
-e .
